requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.10.0+cpu
fastapi==0.135.1
uvicorn==0.42.0
chromadb==1.5.5
sentence-transformers==5.3.0
onnxruntime==1.24.3
pandas==3.0.1
pypdf==6.9.1
python-multipart==0.0.22
python-dotenv